experiments/genai_devrel_weekly_challenge/week1/pyproject.toml
[project]
name = "week1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-auth>=2.38.0",
    "google-auth-oauthlib>=1.2.1",
    "gradio>=5.14.0",
    "pydantic-settings[yaml]>=2.7.1",
]
